// src/app/models/Order.ts
export interface Order {
    id: any; 
    orderCode: any;
    userId: any;
    customerName: any;
    phone: any;
    address: any;
    note: any;
    shippingFee: any; 
    originalPrice: any;
    netPriceTotal: any;
    orderStatusId: any;
    orderStatusName: any;
    paymentStatus: any;
    paymentTypeId: any;
    paymentTypeName: any;
}